refactor: extract SVG export logic to separate module
