+++ plan.md
@@ -1,0 +1,3 @@
+1. Run the full test suite using `pnpm run test` to verify the changes.
+2. Complete pre-commit steps to ensure proper testing, verification, review, and reflection are done.
+3. Submit the change.
